refactor: extract per-format code into row/columnar modules
Moves all feature-specific helpers (encoding, avro serialization, request
construction, log summaries, format-specific tests) from per-callsite cfg
blocks in shared files into two dedicated sibling modules:

  src/row.rs       — compiled when `columnar` is disabled
  src/columnar.rs  — compiled when `columnar` is enabled

`lib.rs` aliases the active module as `crate::format`, so all other files
(`consumer.rs`, `stream.rs`, `listener.rs`) call `crate::format::*`
without any cfg annotations. Each format module exposes the same
internal API:

  build_write_request   — turn drained SeriesBuffer into a StreamWriteRequest
  encode_for_core       — serialize + wrap in an HTTP WriteRequest
  avro_append           — write a request to an avro file
  request_summary       — short human-readable payload summary
  test_support::*       — helpers the stream tests use to assert on payloads

Removes the duplicated assertion blocks in `lib.rs` stream tests in favor
of `format::test_support::count_points_by_channel` and
`first_request_first_double_count`; assertions now read the aggregate
count correctly (5 × 1000 points, not the accidental last-series 1000).

Result: 55 cfg annotations → 8, concentrated at the two boundary points
(module decls in lib.rs + the StreamWriteRequest alias in client.rs).
consumer.rs, stream.rs, listener.rs carry zero cfg. Shared trait signatures
and consumer-wrapper types (Fallback/Dual/Listening) remain unified via
the StreamWriteRequest alias.

Both feature states compile clean under clippy -D warnings and all 6
tests pass in each.
